requirements.txt
pandas
pydantic
pydantic-settings
numpy
pycot-reports
requests